Fix compilation

+++ src/sentry/processing/view_hierarchy_processor.cpp
@@ -31,7 +31,7 @@ Ref<SentryEvent> ViewHierarchyProcessor::process_event(const Ref<SentryEvent> &p
 	if (f) {
 		size_t written = std::fwrite(json_buffer.ptr(), 1, json_buffer.get_used(), f);
 		if (written != json_buffer.get_used()) {
-			sentry::util::print_error(vformat("Failed to write scene tree data - only wrote %d bytes out of %d", written, json_buffer.get_used()));
+			sentry::util::print_error(vformat("Failed to write scene tree data - only wrote %d bytes out of %d", (int64_t)written, (int64_t)json_buffer.get_used()));
 		}
 		std::fclose(f);
 	} else {
